donde-comemos-be/DC/requirements-dev.txt
Django==2.2.2
django-guardian==1.5.1
mysqlclient==1.4.2.post1
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
django-material-admin==1.6.34
djangorestframework==3.9.4
whitenoise
django-cors-headers==3.0.2
coverage==4.5.3
fcm-django==0.2.21
requests==2.22.0

